Bug Fix
Fixed bug where there would sometimes appear several explosions per cow.

+++ Assets/Scripts/CowDeleter.cs
@@ -8,7 +8,7 @@ public class CowDeleter : MonoBehaviour
 {
    private void OnCollisionEnter2D(Collision2D collision)
   {
-    if(collision.gameObject.tag == "RecycleBin")
+    if(collision.gameObject.tag == "RecycleBin" && gameObject.tag != "DeletedCow")
     {
       DeleteCow();
     }
